requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
langchain-huggingface>=0.1.0
langchain-qdrant>=0.2.0
langgraph>=0.2.0
langsmith>=0.1.0
qdrant-client>=1.11.0
sentence-transformers>=3.0.0
fastembed>=0.4.0
pymupdf>=1.24.0
pymupdf4llm>=0.0.16
gradio>=5.0.0
portalocker
python-dotenv